feat: Update manifest.json with default_icon property
This commit updates the manifest.json file to include the "default_icon" property, which specifies the icon to be used as the default popup icon for the extension. This change ensures that the extension has a visually appealing and recognizable icon when the popup is displayed.

Note: The commit message follows the conventional format of starting with a type (feat) followed by a colon and a brief description of the changes.